Diseño componentes en el dashboard del Administrador.

+++ src/app/administrador/navbar/navbar.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { AuthService } from 'src/app/services/auth.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-navbar',
@@ -17,9 +18,29 @@ export class NavbarComponent implements OnInit{
   ngOnInit(): void {
   }
 
-  cerrarSesion() {
-    localStorage.clear();
-    this.router.navigate(['/login']);
+  salir() {
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Con esta acción saldrás del sistema!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Aceptar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        Swal.fire(
+          'Hasta luego!',
+          'Regresa pronto :).',
+          'success'
+        );
+        this.router.navigateByUrl('/login');
+        console.log('Saliendo...');
+      } else {
+        console.log('No hemos salido.');
+      }
+    });
   }
 
 }
